ListField: removed default `None` value for the `items_type` parameter.
This will create more consistent code since `item_types` _is_ required.
Tests had to be modified a bit. Specifically tests which used Model
class with circular reference to itself. It was easier to just extract
the dummy class outside the function, in order to avoid import path hell.

This change breaks the current API, so if merged it should be into a new
major version.