issue #34: default value support in generated schema
In the previous commit we introduced the support for default values
at the BaseField level. That is - fields can have a default value,
which is respected by models.Base.
It did not change the generated schema in any way though.
This commit add support for default values at the schema level.

The "default" keyword is added to the schema only if the user provided
a specific default value for the field. Otherwise the schema is unchanged.

A new test has been written to verify the correctness of the "default"
values in the schema itself.