Optionally disable validators for fields.
JsonModels are heavily used in testing. And for testing purposes
we need to be able to disable type validation in process of filling model with data.
This why I suggest to add extra option to be able to disable validation for fields.
By default validation is on, it will not break current behavior.